Added code to index.js to allow the server to run on port 5000 if a port is not specified. Corrected file paths for files who have had their file path changed within the refactor. Added comments to clarify purpose of some code.
